Improve instructions on NoAccessScreen and bday dialog
