+++ src/ageAssurance/components/NoAccessScreen.tsx
@@ -9,12 +9,11 @@ import {
   useCreateSupportLink,
 } from '#/lib/hooks/useCreateSupportLink'
 import {dateDiff, useGetTimeAgo} from '#/lib/hooks/useTimeAgo'
-import {isAppPassword} from '#/lib/jwt'
 import {logger} from '#/logger'
 import {isWeb} from '#/platform/detection'
 import {isNative} from '#/platform/detection'
 import {useIsBirthdateUpdateAllowed} from '#/state/birthdate'
-import {useSession, useSessionApi} from '#/state/session'
+import {useSessionApi} from '#/state/session'
 import {atoms as a, useBreakpoints, useTheme, web} from '#/alf'
 import {Admonition} from '#/components/Admonition'
 import {AgeAssuranceAppealDialog} from '#/components/ageAssurance/AgeAssuranceAppealDialog'
@@ -30,7 +29,7 @@ import {ShieldCheck_Stroke2_Corner0_Rounded as ShieldIcon} from '#/components/ic
 import {createStaticClick, SimpleInlineLinkText} from '#/components/Link'
 import {Outlet as PortalOutlet} from '#/components/Portal'
 import * as Toast from '#/components/Toast'
-import {Span, Text} from '#/components/Typography'
+import {Text} from '#/components/Typography'
 import {BottomSheetOutlet} from '#/../modules/bottom-sheet'
 import {useAgeAssurance} from '#/ageAssurance'
 import {useAgeAssuranceDataContext} from '#/ageAssurance/data'
@@ -54,9 +53,6 @@ export function NoAccessScreen() {
   const isBirthdateUpdateAllowed = useIsBirthdateUpdateAllowed()
   const {logoutCurrentAccount} = useSessionApi()
   const createSupportLink = useCreateSupportLink()
-
-  const {currentAccount} = useSession()
-  const isUsingAppPassword = isAppPassword(currentAccount?.accessJwt || '')
 
   const aa = useAgeAssurance()
   const isBlocked = aa.state.status === aa.Status.Blocked
@@ -89,22 +85,38 @@ export function NoAccessScreen() {
     logoutCurrentAccount('AgeAssuranceNoAccessScreen')
   }, [logoutCurrentAccount])
 
-  const birthdateUpdateText = canUpdateBirthday ? (
-    <Text style={[textStyles]}>
+  const orgAdmonition = (
+    <Admonition type="tip">
       <Trans>
-        If you believe your birthdate is incorrect, you can update it by{' '}
-        <SimpleInlineLinkText
-          label={_(msg`Click here to update your birthdate`)}
-          style={[textStyles]}
-          {...createStaticClick(() => {
-            logger.metric('ageAssurance:noAccessScreen:openBirthdateDialog', {})
-            birthdateControl.open()
-          })}>
-          clicking here
-        </SimpleInlineLinkText>
-        .
+        For organizational accounts, use the birthdate of the person who is
+        responsible for the account.
       </Trans>
-    </Text>
+    </Admonition>
+  )
+
+  const birthdateUpdateText = canUpdateBirthday ? (
+    <>
+      <Text style={[textStyles]}>
+        <Trans>
+          If you believe your birthdate is incorrect, you can update it by{' '}
+          <SimpleInlineLinkText
+            label={_(msg`Click here to update your birthdate`)}
+            style={[textStyles]}
+            {...createStaticClick(() => {
+              logger.metric(
+                'ageAssurance:noAccessScreen:openBirthdateDialog',
+                {},
+              )
+              birthdateControl.open()
+            })}>
+            clicking here
+          </SimpleInlineLinkText>
+          .
+        </Trans>
+      </Text>
+
+      {orgAdmonition}
+    </>
   ) : (
     <Text style={[textStyles]}>
       <Trans>
@@ -118,15 +130,6 @@ export function NoAccessScreen() {
         .
       </Trans>
     </Text>
-  )
-
-  const orgAdmonition = (
-    <Admonition type="tip">
-      <Trans>
-        For organizational accounts, use the birthdate of the person who is
-        responsible for the account.
-      </Trans>
-    </Admonition>
   )
 
   return (
@@ -175,8 +178,6 @@ export function NoAccessScreen() {
                       </Text>
 
                       {!isBlocked && birthdateUpdateText}
-
-                      {orgAdmonition}
                     </View>
 
                     <AccessSection />
@@ -191,8 +192,6 @@ export function NoAccessScreen() {
                     </Text>
 
                     {birthdateUpdateText}
-
-                    {orgAdmonition}
                   </View>
                 )}
               </>
@@ -224,18 +223,7 @@ export function NoAccessScreen() {
                   </ButtonText>
                 </Button>
 
-                {isUsingAppPassword ? (
-                  <Admonition type="info">
-                    <Trans>
-                      Hmm, it looks like you're logged in with an{' '}
-                      <Span style={[a.italic]}>App Password</Span>. To set your
-                      birthdate, you'll need to log in with your main account
-                      password, or ask whomever controls this account to do so.
-                    </Trans>
-                  </Admonition>
-                ) : (
-                  orgAdmonition
-                )}
+                {orgAdmonition}
               </View>
             )}
